LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_pointercapture-not-lost-in-chorded-buttons-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Pointer Events Capture Test - capture should not be lost early

Test Description: This test checks if setCapture/pointerup functions works properly. Complete the following actions:
Put your mouse over the red box
Press and hold left mouse button. Box will call setPointerCapture
Press right button and release
Pointer capture should not be lost
Press right button again and release
Pointer capture should not be lost
Release left mouse button. lostpointercapture is called

Test passes if the proper behavior of the events is observed.

Harness Error (FAIL), message = Unhandled rejection: unimplemented

TIMEOUT Pointer Events Capture Test - capture not lost due to chorded buttons interaction Test timed out
NOTRUN Pointer Events Capture Test - capture not lost due to combination of left and right chorded buttons interaction.
